+++ modules/Food_Service/includes/DeleteTransactionItem.fnc.php
@@ -15,15 +15,63 @@ function DeleteTransactionItem( $transaction_id, $item_id, $type = 'student' )
 {
 	if ( $type == 'staff' )
 	{
-		$sql1 = "UPDATE food_service_staff_transactions SET BALANCE=BALANCE-(SELECT AMOUNT FROM food_service_staff_transaction_items WHERE TRANSACTION_ID='" . (int) $transaction_id . "' AND ITEM_ID='" . (int) $item_id . "') WHERE TRANSACTION_ID>='" . $transaction_id . "' AND STAFF_ID=(SELECT STAFF_ID FROM food_service_staff_transactions WHERE TRANSACTION_ID='" . (int) $transaction_id . "')";
-		$sql2 = "UPDATE food_service_staff_accounts SET BALANCE=BALANCE-(SELECT AMOUNT FROM food_service_staff_transaction_items WHERE TRANSACTION_ID='" . (int) $transaction_id . "' AND ITEM_ID='" . (int) $item_id . "') WHERE STAFF_ID=(SELECT STAFF_ID FROM food_service_staff_transactions WHERE TRANSACTION_ID='" . (int) $transaction_id . "')";
-		$sql3 = "DELETE FROM food_service_staff_transaction_items WHERE TRANSACTION_ID='" . (int) $transaction_id . "' AND ITEM_ID='" . (int) $item_id . "'";
+		/**
+		 * Fix MySQL 5.6 error Can't specify target table for update in FROM clause.
+		 *
+		 * @link https://stackoverflow.com/questions/45494/mysql-error-1093-cant-specify-target-table-for-update-in-from-clause
+		 */
+		$staff_id = DBGetOne( "SELECT STAFF_ID
+			FROM food_service_staff_transactions
+			WHERE TRANSACTION_ID='" . (int) $transaction_id . "'" );
+
+		$sql1 = "UPDATE food_service_staff_transactions
+		SET BALANCE=BALANCE-(SELECT AMOUNT
+			FROM food_service_staff_transaction_items
+			WHERE TRANSACTION_ID='" . (int) $transaction_id . "'
+			AND ITEM_ID='" . (int) $item_id . "')
+		WHERE TRANSACTION_ID>='" . (int) $transaction_id . "'
+		AND STAFF_ID='" . (int) $staff_id . "'";
+
+		$sql2 = "UPDATE food_service_staff_accounts
+		SET BALANCE=BALANCE-(SELECT AMOUNT
+			FROM food_service_staff_transaction_items
+			WHERE TRANSACTION_ID='" . (int) $transaction_id . "'
+			AND ITEM_ID='" . (int) $item_id . "')
+		WHERE STAFF_ID='" . (int) $staff_id . "'";
+
+		$sql3 = "DELETE FROM food_service_staff_transaction_items
+			WHERE TRANSACTION_ID='" . (int) $transaction_id . "'
+			AND ITEM_ID='" . (int) $item_id . "'";
 	}
 	else
 	{
-		$sql1 = "UPDATE food_service_transactions SET BALANCE=BALANCE-(SELECT AMOUNT FROM food_service_transaction_items WHERE TRANSACTION_ID='" . (int) $transaction_id . "' AND ITEM_ID='" . (int) $item_id . "') WHERE TRANSACTION_ID>='" . $transaction_id . "' AND ACCOUNT_ID=(SELECT ACCOUNT_ID FROM food_service_transactions WHERE TRANSACTION_ID='" . (int) $transaction_id . "')";
-		$sql2 = "UPDATE food_service_accounts SET BALANCE=BALANCE-(SELECT AMOUNT FROM food_service_transaction_items WHERE TRANSACTION_ID='" . (int) $transaction_id . "' AND ITEM_ID='" . (int) $item_id . "') WHERE ACCOUNT_ID=(SELECT ACCOUNT_ID FROM food_service_transactions WHERE TRANSACTION_ID='" . (int) $transaction_id . "')";
-		$sql3 = "DELETE FROM food_service_transaction_items WHERE TRANSACTION_ID='" . (int) $transaction_id . "' AND ITEM_ID='" . (int) $item_id . "'";
+		/**
+		 * Fix MySQL 5.6 error Can't specify target table for update in FROM clause.
+		 *
+		 * @link https://stackoverflow.com/questions/45494/mysql-error-1093-cant-specify-target-table-for-update-in-from-clause
+		 */
+		$account_id = DBGetOne( "SELECT ACCOUNT_ID
+			FROM food_service_transactions
+			WHERE TRANSACTION_ID='" . (int) $transaction_id . "'" );
+
+		$sql1 = "UPDATE food_service_transactions
+		SET BALANCE=BALANCE-(SELECT AMOUNT
+			FROM food_service_transaction_items
+			WHERE TRANSACTION_ID='" . (int) $transaction_id . "'
+			AND ITEM_ID='" . (int) $item_id . "')
+		WHERE TRANSACTION_ID>='" . (int) $transaction_id . "'
+		AND ACCOUNT_ID='" . (int) $account_id . "'";
+
+		$sql2 = "UPDATE food_service_accounts
+		SET BALANCE=BALANCE-(SELECT AMOUNT
+			FROM food_service_transaction_items
+			WHERE TRANSACTION_ID='" . (int) $transaction_id . "'
+			ND ITEM_ID='" . (int) $item_id . "')
+		WHERE ACCOUNT_ID='" . (int) $account_id . "'";
+
+		$sql3 = "DELETE FROM food_service_transaction_items
+			WHERE TRANSACTION_ID='" . (int) $transaction_id . "'
+			AND ITEM_ID='" . (int) $item_id . "'";
 	}
 
 	db_trans_start();