+++ CHANGES.md
@@ -106,6 +106,7 @@ Changes in 10.0
 - SQL use GROUP BY instead of DISCTINCT ON for MySQL compatibility in Address.inc.php & EnterEligibility.php
 - SQL cast Config( 'STUDENTS_EMAIL_FIELD' ) to int when custom field in SendNotification.fnc.php, Registration.fnc.php, Moodle/getconfig.inc.php & ImportUsers.fnc.php
 - Fix MySQL 5.6 syntax error when WHERE without FROM clause in TakeAttendance.php & InputFinalGrades.php
+- Fix MySQL 5.6 error Can't specify target table for update in FROM clause in PortalPollsNotes.fnc.php, DeleteTransaction.fnc.php, DeleteTransactionItem.fnc.php, Rollover.php, CopySchool.php & AssignOtherInfo.php
 
 Changes in 9.3.1
 ----------------